nta-flink-stream/src/main/resources/script/requirements.txt
kafka-python~=2.2.15
pytz~=2024.2
schedule~=1.2.2
paramiko~=3.5.1